Better collision detection

+++ app/src/main/java/com/example/brickbreaker/data/Ball.kt
@@ -4,7 +4,7 @@ data class Ball(
     // position on the canvas
     val position: Pair<Float, Float> = Pair(300F, 1000F),
     // radius
-    val radius: Int = 15,
+    val radius: Float = 15F,
     // pixels per second
     val velocity: Pair<Float, Float> = Pair(6F, 6F)
 )
